Added unit tests around config and other minor tweaks.

+++ src/main/java/com/opencredo/connect/venafi/tpp/log/api/TppLog.java
@@ -23,7 +23,7 @@ public interface TppLog {
 
     //If we in the future need to send a query Param with +
     // be aware TPP server decodes plus as space so we'd have to actively encode + as %2B
-    default LogResponse getLogs(String token, String fromTime, String limit, int offset) {
+    default LogResponse getLogs(String token, String fromTime, String limit, long offset) {
         Map<String, Object> queryParams = new HashMap<>();
         queryParams.put(FROM_TIME, fromTime);
         queryParams.put(LIMIT, limit);
